Split InternalVersions* into one file per type
